requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astroid=2.4.2=py37_0
certifi=2020.6.20=py37_0
colorama=0.4.3=py_0
intel-openmp=2020.1=216
isort=4.3.21=py37_0
joblib=0.16.0=py_0
lazy-object-proxy=1.4.3=py37he774522_0
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
liblapack=3.8.0=16_mkl
mccabe=0.6.1=py37_1
mkl=2020.1=216
numpy=1.19.1=py37hae9e721_0
pandas=1.0.5=py37h47e9c7a_0
pip=20.1.1=py37_1
pylint=2.5.3=py37_0
python=3.7.1=h33f27b4_4
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
scikit-surprise=1.1.1=py37h44b1f71_0
setuptools=49.2.0=py37_0
six=1.15.0=pyh9f0ad1d_0
toml=0.10.1=py_0
typed-ast=1.4.1=py37he774522_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
